move startStreamingLog to SpringCloudDeployment
