+++ azure-toolkit-libs/azure-toolkit-springcloud-lib/src/main/java/com/microsoft/azure/toolkit/lib/springcloud/task/DeploySpringCloudAppTask.java
@@ -9,12 +9,9 @@ import com.microsoft.azure.toolkit.lib.Azure;
 import com.microsoft.azure.toolkit.lib.common.bundle.AzureString;
 import com.microsoft.azure.toolkit.lib.common.exception.AzureToolkitRuntimeException;
 import com.microsoft.azure.toolkit.lib.common.messager.AzureMessager;
-import com.microsoft.azure.toolkit.lib.common.messager.IAzureMessager;
 import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
 import com.microsoft.azure.toolkit.lib.common.operation.OperationContext;
 import com.microsoft.azure.toolkit.lib.common.task.AzureTask;
-import com.microsoft.azure.toolkit.lib.common.utils.Debouncer;
-import com.microsoft.azure.toolkit.lib.common.utils.TailingDebouncer;
 import com.microsoft.azure.toolkit.lib.springcloud.AzureSpringCloud;
 import com.microsoft.azure.toolkit.lib.springcloud.SpringCloudAppDraft;
 import com.microsoft.azure.toolkit.lib.springcloud.SpringCloudCluster;
@@ -25,15 +22,12 @@ import com.microsoft.azure.toolkit.lib.springcloud.config.SpringCloudDeploymentC
 import lombok.Getter;
 import org.apache.commons.lang3.StringUtils;
 import reactor.core.Disposable;
-import reactor.core.scheduler.Schedulers;
 
 import javax.annotation.Nonnull;
-import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.concurrent.CountDownLatch;
 
 @Getter
 public class DeploySpringCloudAppTask extends AzureTask<SpringCloudDeployment> {
@@ -107,7 +101,6 @@ public class DeploySpringCloudAppTask extends AzureTask<SpringCloudDeployment> {
             } catch (final Exception e) {
                 app.refresh();
                 this.deployment = app.getActiveDeployment();
-                startStreamingLog(true);
                 throw new AzureToolkitRuntimeException(e);
             }
         }));
@@ -133,50 +126,13 @@ public class DeploySpringCloudAppTask extends AzureTask<SpringCloudDeployment> {
         return this.deployment;
     }
 
-    private void startStreamingLog(boolean follow) {
-        if (Objects.isNull(this.deployment) || !openStreamingLogOnFailure) {
-            return;
-        }
-        final IAzureMessager messager = AzureMessager.getMessager();
-        Optional.ofNullable(this.deployment.getLatestInstance()).ifPresent(i -> {
-            messager.info(AzureString.format("Opening streaming log of instance({0})...", i.getName()));
-            messager.debug("###############STREAMING LOG BEGIN##################");
-            final CountDownLatch latch = new CountDownLatch(1);
-            final Debouncer fireEvents = new TailingDebouncer(() -> {
-                stopStreamingLog();
-                latch.countDown();
-            }, Long.valueOf(Duration.ofSeconds(15).toMillis()).intValue());
-            // refer to https://github.com/Azure/azure-cli-extensions/blob/main/src/spring/azext_spring/app.py#app_tail_log_internal
-            final SpringCloudCluster service = deployment.getParent().getParent();
-            disposable = this.deployment.getLatestInstance().streamingLogs(follow, service.isConsumptionTier() ? 300 : 500)
-                    .doFinally(type -> messager.debug("###############STREAMING LOG END##################"))
-                    .subscribeOn(Schedulers.boundedElastic())
-                    .subscribe((s) -> {
-                        messager.debug(s);
-                        fireEvents.debounce();
-                    });
-            fireEvents.debounce();
-            try {
-                latch.await();
-            } catch (final InterruptedException e) {
-                stopStreamingLog();
-            } finally {
-                latch.countDown();
-            }
-        });
-    }
-
-    private void stopStreamingLog() {
-        if (!disposable.isDisposed()) {
-            disposable.dispose();
-        }
-    }
-
     private void startApp() {
         AzureMessager.getMessager().info(START_APP);
-        if (!deployment.waitUntilReady(TIMEOUT_IN_SECONDS)) {
+        if (!this.deployment.waitUntilReady(TIMEOUT_IN_SECONDS)) {
             AzureMessager.getMessager().warning(GET_APP_STATUS_TIMEOUT);
-            startStreamingLog(false);
+            if (Objects.nonNull(this.deployment) && openStreamingLogOnFailure) {
+                this.deployment.startStreamingLog(false);
+            }
         }
     }
 }
